refactor: split component to different file
