+++ app/pages/[username]/[slug]/index.vue
@@ -4,49 +4,42 @@ import id from 'date-fns/locale/id'
 
 defineOgImageComponent('default')
 
-const user = useSupabaseUser()
 const supabase = useSupabaseClient()
 const route = useRoute()
-const toast = useToast()
-const openLoginModal = ref(false)
-const openShareModal = ref(false)
 const slug = route.params.slug
 const authorUsername = route.params.username.toString()
 
-const { data: story, refresh } = await useAsyncData(
-  `story/${slug}`,
-  async () => {
-    const { data, error } = await supabase
-      .from('stories')
-      .select(
-        `*,
+const { data: story } = await useAsyncData(`story/${slug}`, async () => {
+  const { data, error } = await supabase
+    .from('stories')
+    .select(
+      `*,
       tags:story_tags!id(tag:tag_id(slug)),
       author:users(id, bio, display_name, username, location, created_at),
       reactions:story_reactions!id(*)
       `,
-      )
-      .eq('slug', slug.toString())
-      .single()
+    )
+    .eq('slug', slug.toString())
+    .single()
 
-    if (error) {
-      console.error(error)
-      return null
-    }
+  if (error) {
+    console.error(error)
+    return null
+  }
 
-    const coverWithPath = data.cover_path
-      ? supabase.storage.from('story-cover').getPublicUrl(data.cover_path).data
-          .publicUrl
-      : null
+  const coverWithPath = data.cover_path
+    ? supabase.storage.from('story-cover').getPublicUrl(data.cover_path).data
+        .publicUrl
+    : null
 
-    return {
-      ...data,
-      cover_path: coverWithPath,
-      tags: data.tags.map(
-        tag => (tag.tag as unknown as { slug: string }).slug as string,
-      ),
-    }
-  },
-)
+  return {
+    ...data,
+    cover_path: coverWithPath,
+    tags: data.tags.map(
+      tag => (tag.tag as unknown as { slug: string }).slug as string,
+    ),
+  }
+})
 
 if (!story.value) {
   throw createError({
@@ -69,55 +62,6 @@ const breadcrumbs = [
   },
 ]
 
-const isUserHasReacted = computed(() => {
-  const userId = user.value?.id
-
-  if (!userId || !story.value) {
-    return false
-  }
-
-  const hasReacted = story.value.reactions.some(
-    react => react.reacted_by === userId,
-  )
-
-  return hasReacted
-})
-
-const likeStory = async () => {
-  if (!user.value) {
-    openLoginModal.value = true
-    return
-  }
-
-  const userId = user.value.id
-
-  try {
-    if (isUserHasReacted.value) {
-      await supabase.from('story_reactions').delete().eq('reacted_by', userId)
-    } else {
-      await supabase
-        .from('story_reactions')
-        .insert({ story: story.value.id, reacted_by: userId })
-    }
-  } catch (error) {
-    console.error(error)
-  } finally {
-    refresh()
-  }
-}
-
-const shareStory = () => {
-  openShareModal.value = true
-}
-
-provide(onSuccessLogin, () => {
-  openLoginModal.value = false
-  toast.add({
-    title: 'Berhasil login',
-    color: 'success',
-  })
-})
-
 useSeoMeta({
   title: story.value?.title,
   description: `Cerita berjudul ${story.value.title} dari penulis ${story.value.author.display_name}`,
@@ -131,31 +75,7 @@ useSeoMeta({
     </div>
 
     <div class="mt-4 flex w-full flex-col-reverse gap-4 lg:flex-row">
-      <div class="mt-8 flex w-24 items-center gap-2 px-4 md:flex-col md:px-0">
-        <UTooltip text="Sukai cerita">
-          <UButton
-            icon="i-heroicons-hand-thumb-up-solid"
-            variant="soft"
-            :color="isUserHasReacted ? 'primary' : 'neutral'"
-            class="flex md:flex-col"
-            @click="likeStory"
-          >
-            <span class="block">
-              {{ story.reactions.length }}
-            </span>
-          </UButton>
-        </UTooltip>
-        <UTooltip text="Bagikan cerita">
-          <UButton
-            icon="heroicons:share"
-            variant="soft"
-            color="neutral"
-            aria-label="Button to share this article"
-            @click="shareStory"
-          />
-        </UTooltip>
-      </div>
-
+      <StoryReactions :story="story" />
       <div class="flex w-full flex-col">
         <div
           class="w-full overflow-hidden shadow md:rounded-lg md:border md:border-neutral-300 md:bg-neutral-50"
@@ -228,44 +148,7 @@ useSeoMeta({
         </SharedJoinBanner>
       </div>
 
-      <div
-        class="hidden h-full w-full rounded-lg border border-neutral-300 bg-neutral-50 p-4 shadow md:block lg:max-w-[370px]"
-      >
-        <NuxtLink
-          :to="`/${authorUsername}`"
-          class="group mb-4 flex items-center gap-4 rounded"
-          title="To author page"
-        >
-          <UserPicture :seed="authorUsername" width="40" height="40" />
-          <b class="group-hover:text-primary-600">{{
-            story.author.display_name
-          }}</b>
-        </NuxtLink>
-
-        <p class="text-neutral-600">
-          {{ story.author.bio }}
-        </p>
-        <ul class="mt-4 space-y-4 text-sm">
-          <li v-if="story.author.location">
-            <b class="text-xs">DOMISILI</b>
-            <p class="text-neutral-600">{{ story.author.location }}</p>
-          </li>
-          <li>
-            <b class="text-xs">BERGABUNG</b>
-            <p class="text-neutral-600">
-              {{
-                format(new Date(story.author.created_at), 'PPP', { locale: id })
-              }}
-            </p>
-          </li>
-        </ul>
-      </div>
+      <StoryAside :author="story.author" />
     </div>
-    <LazySharedLoginModal v-model:open="openLoginModal" />
-    <LazySharedShareStoryModal
-      v-model:open="openShareModal"
-      :title="story.title"
-      :author="story.author.display_name"
-    />
   </div>
 </template>
