+++ app/types/entities.ts
@@ -8,3 +8,5 @@ export type StoryStatusHistory =
   Database['public']['Tables']['story_status_histories']['Row']
 export type ReadingHabit = Database['public']['Tables']['reading_habits']['Row']
 export type BookGenre = Database['public']['Tables']['book_genres']['Row']
+export type StoryReaction =
+  Database['public']['Tables']['story_reactions']['Row']
